Tests/LibWeb/Text/expected/css/has-fast-reject-filter-sibling-combinator-invalidation.txt

initial match: false
match after adding sibling: true
